backend/requirements.txt
    # Smart School Backend Dependencies
    # Core Flask dependencies
    Flask==2.3.3
    Flask-CORS==4.0.0
    Flask-SQLAlchemy==3.1.1
    Flask-Migrate==4.0.5
    Flask-JWT-Extended==4.5.3
    python-dotenv==1.0.0

    # Database
    PyMySQL==1.1.0
    SQLAlchemy==2.0.36

    # Data processing
    numpy==1.24.3
    pandas==2.0.3
    scikit-learn==1.3.0

    # HTTP requests
    requests==2.31.0

# Additional utilities
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2

# App-specific
passlib==1.7.4
httpx==0.27.2